pyproject.toml
[project]
name = "libary-usage-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro Vieira",email = "pedro515vieira@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "api-crud-generate-libary @ git+https://github.com/pedro-vlps/api_crud_generate_libary.git",
    "fastapi (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlacodegen (>=4.0.3,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
